+++ txcdr-client/app/(auth)/event/VolRegisteredPage.tsx
@@ -1,0 +1,5 @@
+import { PageProps } from "./types";
+
+export function VolRegisteredPage(props: PageProps) {
+  return <></>;
+}
